LPD-32543 SF, add TypeAlias to support the new feature
# breaking

## What modules/core/petra/petra-sql-dsl-api/src/main/java/com/liferay/petra/sql/dsl/expression/Expression.java

Add new method Expression#as(String, Class<?>)

## Why

This is needed to set specific return type for alias

----
