+++ modules/core/petra/petra-sql-dsl-api/src/main/java/com/liferay/petra/sql/dsl/expression/Alias.java
@@ -12,8 +12,6 @@ public interface Alias<T> extends Expression<T> {
 
 	public Expression<T> getExpression();
 
-	public Class<?> getJavaType();
-
 	public String getName();
 
 }
