+++ modules/core/petra/petra-sql-dsl-api/src/main/java/com/liferay/petra/sql/dsl/expression/Expression.java
@@ -14,9 +14,9 @@ import com.liferay.petra.sql.dsl.query.sort.OrderByExpression;
  */
 public interface Expression<T> extends ASTNode {
 
-	public Alias<T> as(Class<?> javaType, String name);
-
 	public Alias<T> as(String name);
+
+	public Alias<T> as(String name, Class<?> javaType);
 
 	public OrderByExpression ascending();
