+++ modules/core/petra/petra-sql-dsl-api/src/main/java/com/liferay/petra/sql/dsl/expression/TypeAlias.java
@@ -1,0 +1,15 @@
+/**
+ * SPDX-FileCopyrightText: (c) 2024 Liferay, Inc. https://liferay.com
+ * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
+ */
+
+package com.liferay.petra.sql.dsl.expression;
+
+/**
+ * @author Tina Tian
+ */
+public interface TypeAlias<T> extends Alias<T> {
+
+	public Class<?> getJavaType();
+
+}
