+++ modules/core/petra/petra-sql-dsl-spi/src/main/java/com/liferay/petra/sql/dsl/spi/expression/DefaultExpression.java
@@ -19,13 +19,13 @@ import com.liferay.petra.sql.dsl.spi.query.sort.DefaultOrderByExpression;
 public interface DefaultExpression<T> extends Expression<T> {
 
 	@Override
-	public default Alias<T> as(Class<?> javaType, String name) {
-		return new DefaultAlias<>(this, javaType, name);
+	public default Alias<T> as(String name) {
+		return new DefaultAlias<>(this, name);
 	}
 
 	@Override
-	public default Alias<T> as(String name) {
-		return new DefaultAlias<>(this, name);
+	public default Alias<T> as(String name, Class<?> javaType) {
+		return new DefaultTypeAlias<>(this, name, javaType);
 	}
 
 	@Override
